Refacto to split types
